+++ server/games-service/src/main/java/com/fnd/games_store/games/controller/SpecificGameController.java
@@ -1,6 +1,6 @@
 package com.fnd.games_store.games.controller;
 
-import com.fnd.games_store.games.dto.GameResponseDTO;
+import com.fnd.games_store.games.dto.game.GameResponseDTO;
 import org.springframework.http.ResponseEntity;
 
 public interface SpecificGameController {
