+++ server/games-service/src/main/java/com/fnd/games_store/games/controller/implementation/SpecificGameControllerImpl.java
@@ -1,7 +1,7 @@
 package com.fnd.games_store.games.controller.implementation;
 
 import com.fnd.games_store.games.controller.SpecificGameController;
-import com.fnd.games_store.games.dto.GameResponseDTO;
+import com.fnd.games_store.games.dto.game.GameResponseDTO;
 import com.fnd.games_store.games.service.SpecificGameService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
