+++ server/games-service/src/main/java/com/fnd/games_store/games/controller/implementation/SpecificGameListControllerImpl.java
@@ -2,7 +2,7 @@ package com.fnd.games_store.games.controller.implementation;
 
 
 import com.fnd.games_store.games.controller.SpecificGameListController;
-import com.fnd.games_store.games.dto.GameResponseDTO;
+import com.fnd.games_store.games.dto.game.GameResponseDTO;
 import com.fnd.games_store.games.service.SpecificGameListService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
