+++ server/games-service/src/main/java/com/fnd/games_store/games/dto/game/GameRequestDTO.java
@@ -1,4 +1,4 @@
-package com.fnd.games_store.games.dto;
+package com.fnd.games_store.games.dto.game;
 
 import lombok.*;
 import java.math.BigDecimal;
