+++ server/games-service/src/main/java/com/fnd/games_store/games/service/SpecificGameService.java
@@ -1,6 +1,6 @@
 package com.fnd.games_store.games.service;
 
-import com.fnd.games_store.games.dto.GameResponseDTO;
+import com.fnd.games_store.games.dto.game.GameResponseDTO;
 
 public interface SpecificGameService {
