+++ server/games-service/src/main/java/com/fnd/games_store/games/service/implementation/SpecificGameListServiceImpl.java
@@ -1,7 +1,6 @@
 package com.fnd.games_store.games.service.implementation;
 
-import com.fnd.games_store.games.dto.GameResponseDTO;
-import com.fnd.games_store.games.exceptions.GameNotFoundException;
+import com.fnd.games_store.games.dto.game.GameResponseDTO;
 import com.fnd.games_store.games.repository.GameRepository;
 import com.fnd.games_store.games.service.SpecificGameListService;
 import org.springframework.beans.factory.annotation.Autowired;
