+++ server/games-service/src/main/java/com/fnd/games_store/games/service/implementation/SpecificGameServiceImpl.java
@@ -1,6 +1,6 @@
 package com.fnd.games_store.games.service.implementation;
 
-import com.fnd.games_store.games.dto.GameResponseDTO;
+import com.fnd.games_store.games.dto.game.GameResponseDTO;
 import com.fnd.games_store.games.exceptions.GameNotFoundException;
 import com.fnd.games_store.games.repository.GameRepository;
 import com.fnd.games_store.games.service.SpecificGameService;
