+++ server/games-service/src/test/java/com/fnd/games_store/service/GetGameByName_IntegrationTest.java
@@ -1,58 +1,71 @@
 package com.fnd.games_store.service;
 
 import com.fnd.games_store.games.GamesApplication;
+import com.fnd.games_store.games.dto.game.GameResponseDTO;
+import com.fnd.games_store.games.service.SpecificGameService;
 import lombok.extern.slf4j.Slf4j;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import static org.assertj.core.api.Assertions.*;
 
 @SpringBootTest(classes = GamesApplication.class)
 @Slf4j
+
+
 public class GetGameByName_IntegrationTest {
 
 
-//    @Autowired
-//    private GameRepository repository;
+    @Autowired
+    private SpecificGameService service;
+
+    private GameResponseDTO serviceGameEntity;
+
+    private GameResponseDTO controlGameEntity;
+
+    private String controlGameName = "Grand Theft Auto V";
+
+
+    @Test
+    void getGameByName_ShouldReturnProperGameResponseDTO(){
+        assertThat(service.getGameByName(controlGameName)).isEqualTo(controlGameEntity);
+    }
+
+
+    @BeforeEach
+    void testSetup(){
+
+
+        this.controlGameEntity = createAppropriateGameResponseDTO();
+
+
+    }
+
+    private GameResponseDTO createAppropriateGameResponseDTO() {
+        GameResponseDTO returnableGame = new GameResponseDTO();
+        returnableGame.setId("1");
+        returnableGame.setName("Doom Eternal");
+        returnableGame.setReleaseDate("2019");
+        returnableGame.setDeveloper(null);
+        returnableGame.setPublisher(null);
+
+
+
+        return returnableGame;
+    }
+
+    //    private List<GameResponseDTO> referenceList;
 //
-//    private Game referenceGame;
+//    private Integer page = 0;
 //
-//    private String referenceGameName = "DOOM Eternal";
+//    private Integer pageSize = 2;
 //
-//
-//    @Test
-//    void getGameByName_ShouldReturnProperGameResponseDTO(){
-//
-//
-////        Session session = Session.getSession();
-//
-//
-//        assertThat(referenceGame).isEqualTo(createAppropriateGameResponseDTO());
-//    }
-//
-//
-//    private Game createAppropriateGameResponseDTO() {
-//        Game returnableGame = new Game();
-//        returnableGame.setId("1");
-//        returnableGame.setName("Doom Eternal");
-//        returnableGame.setReleaseDate("2019");
-//        returnableGame.setDeveloper(null);
-//        returnableGame.setPublisher(null);
-//
-//
-//
-////        '1','DOOM Eternal', '2019', '1','1', 1500.64, 15.2,
-//        return returnableGame;
-//    }
-//
-//    //    private List<GameResponseDTO> referenceList;
-////
-////    private Integer page = 0;
-////
-////    private Integer pageSize = 2;
-////
-////    private String sortField = "name";
-////    private Sort sortBy = Sort.by(sortField);
-//
-//    //        referenceList = repository.findAll(PageRequest.of(page,pageSize, sortBy)).stream().map(GameResponseDTO::new).collect(Collectors.toList());
+//    private String sortField = "name";
+//    private Sort sortBy = Sort.by(sortField);
+
+    //        referenceList = repository.findAll(PageRequest.of(page,pageSize, sortBy)).stream().map(GameResponseDTO::new).collect(Collectors.toList());
